+++ changelog.md
@@ -1,6 +1,7 @@
 TGUI 1.10  (TBD)
 ----------------
 
+- Added ClientPadding property to ChildWindow renderer
 - Added option to scale image (stretching while maintaining ratio) before applying 9-slice scaling
 - Show effects didn't take origin into account
 - Changed scaling when object is smaller than corners of 9-slice